Upgrade dataloader to 2.2.2
This crosses a major version; see https://github.com/graphql/dataloader/releases/tag/v2.0.0
for changes. The main change is that the timing of promise resolution
has changed, in a way that avoids N+1 select in some cases. This also
adds Typescript definitions, and changes how error handling works. We
convert calls to loadMany to go through a wrapper, loadByIds, to change
that error handling again to something friendlier.