Improve UI for adding draft posts to sequences
For historical/technical reasons, there's a UI to add posts to
sequences and, beneath it, a different UI to add *draft* posts to
sequences. The UI to add posts allows you to add multiple posts and
rearrange them before you save the new list of posts all at once. The UI
to add draft posts adds the post and saves the list immediately. Because
of this, if you add/rearrange posts in the first UI and don't save yet,
and then you add a draft post, the saving of the draft post uses the old
list, overwriting the changes you previously made.

This commit prevents that, by disabling the the UI to add draft posts
when there are unsaved changes in the first UI.