Fix the account-edit form for cross-user editing
The vulcan-forms refactoring PR
(https://github.com/ForumMagnum/ForumMagnum/pull/6878) broken the
edit-user form when editing users other than yourself (but not when
editing yourself). This is because, when switching from the `withSingle`
HoC to the `useSingle` hook, there was a case that `withSingle` handled
that `useSingle` didn't: when no documentId is passed, but a slug is
passed instead. Tightening the typechecking on both the `useSingle` and
`FormWrapper` components confirmed that `UsersEditForm` and
`EditProfileForm` were affected, but nothing else in the codebase was.

Fix by extending `useSingle` to be able to take a slug.