Fix links in LW legacy format being broken on AF
This issue was reported as: Tag page
  https://www.alignmentforum.org/tag/holden-karnofsky
contains a link to
  https://www.alignmentforum.org/lw/cbs/thoughts_on_the_singularity_institute_si/
which is broken. Actually, the link was to
  http://www.lesswrong.com/lw/cbs/thoughts_on_the_singularity_institute_si/
The PostLinkPreviewLegacy component changed the http->https and the
lesswrong.com->alignmentforum.org, then failed to have a working hover
preview and making a broken link when clicked.

This changes the `legacyIdPost` resolver so it still works even if the
default view contains `af:true` (ie, if it's being used by AF). This
makes the link preview on that tag page work, and makes the
alignmentforum version of the legacy post link function as a redirect to
LW like it should.