Add a bunch of validation rules for usernames
At least 2 characters, can't start or end with whitespace, no
consecutive whitespace characters, and certain special characters are
not allowed. (This is *not* meant to be load-bearing for preventing XSS;
a sloppy third-party tool might be saved from XSS by it, but it's just
meant to prevent making the user name display confusing.)

This is only enforced on password-login users (ie, not on usernames
coming from OAuth providers), and only at account creation time. So
users with preexisting one-character names and such are grandfathered in
and can keep them.