Fix a missing between username and date when copy-pasting out of dialogues
On the theory that what some users want is to have a dialogue, then have
someone take it and edit it, and that what the editor will probably do
is copy-paste the dialogue into the New Post page or Google Docs, I
tried that copy paste. It mostly worked, but usernames and
comment-timestamps appeared next to each other with no space. So, add a
space (and reduce the margin by the corresponding amount.)