Add a "block ownership" mode to the collab editor
This is a prototype for a way of running debates. It is not
production-ready.

This makes a mode for CkEditor5 collaborative-editing documents where
users can have ownership over paragraphs, which prevents other users
from editing those paragraphs. You own a paragraph if (1) you are the
post author or are explicitly shared (not link-shared), and (2) the
paragraph starts with "Yourname:". If someone tries to edit your
paragraph, it will but them in Suggesting mode, and they can't approve
their own suggestions.

A paragraph can be edited by anyone if it doesn't start with a name and
a ':', or it starts with the name of someone who isn't explicitly shared
on the post.

This mode is only active if you put "&blockOwnership=true" in the URL of
the post-edit page, eg

  http://localhost:3000/editPost?postId=POSTIDHERE&blockOwnership=true.

Each participant must do this manually. If a user doesn't have
block-ownership mode turned on or is using a different branch to edit
the post, and they have write access to the post, then they can edit
every block inside the post even ones that are owned by others.

This also adds a button to the bottom of the post editor, labelled "Add
Message". This inserts a paragraph at the end of the document containing
"Username:".