Vendor cluster of libraries around dom-anchor-text-quote
Vendors dom-anchor-text-quote, dom-anchor-text-position, and dom-seek
(all of which are by the same author). The contents of the files aren't
changed relative to what's in github, other than adding a header with
the source link and a license, and adding :AnyBecauseTodo type
annotations.

This fixes a crash when an inline reaction contains the last character
of a comment. The fix comes just from bringing everything up to date
with what's in github (as opposed to what's in npm).