Automatically display new posts on the frontpage (#18)
This PR makes the changes described in the [Update logic to
automatically display new posts on the
frontpage](https://app.clickup.com/t/8685tkcnv) ClickUp task.

* All posts get a frontpageDate value on creation, like on the EA Forum.
(This is achieved via a change to posts/schema.tsx, changing the
onInsert code that used to run just for the EA Forum to run on all
not-LW-nor-AF sites. (Maybe later it'll be a setting, but that'll
require coordination with the other teams.)
* Also, approving a user no longer bumps the postedAt dates of their
posts (this was useful when posts were hidden while users weren't
approved, but we no longer hide unreviewed user posts).
* showFirstPostReviewMessageSetting removes the message about first
posts awaiting moderator approval.
* (The ClickUp item about what happens when the moderators click
Frontpage or Personal Blogpost required no change.)
* (The ClickUp item about removing the karma bonus for appearing on the
frontpage is achieved by a setting change in ForumCredentials.)
* showPersonalBlogpostIconSetting prevents the personal blogpost icon
appearing.
* I've also made the label of the terms of service checkbox clickable.
(This isn't part of the ClickUp task, but is nice to have.)

I'll also make a PR for the settings changes in ForumCredentials.