Fix perpetual profile image loading animation (#45)
As described in [this ClickUp
issue](https://app.clickup.com/t/86864gfkb), there's a bug in the
upstream code where a CSS class that animates a placeholder for loading
images is never removed, so it's still animating behind the image after
it loads. Because the animation was behind the image, it wasn't
noticeable until someone uploaded an image (the Waking Up logo) with
transparent parts, and then parts of the animation became visible.

This PR adds a useState hook to track whether the image is loading, and
only applies the CSS class while it is. I use the React onLoad event in
<img> to update the state. Unfortunately, that event isn't reliable; so
there's also some fallback code in case it doesn't work.