Fix lingering logo focus state; fix oversized header bar at narrow widths (#53)
See the comments in [this ClickUp
task](https://app.clickup.com/t/86864dwnq) for the spec for this PR.

Clicking the logo takes you to the homepage, but it also leaves the
header element with the :focus state. The :focus state had an opacity
0.7, like :hover, so it looked like the hover state doesn't leave if you
move the mouse away after a click. I've removed the rule for the :focus
state.

At very narrow widths (below about 380px), the header bar overflowed the
view width, causing horizontal scrolling. I've fixed this by narrowing
the button padding and making the logo adjust its width to the remaining
space with flex-grow.