Update react-router and react-router-dom from v5 to v6
Typescript types are now included to the @types packages are removed.

The main change is that `useNavigation` has been replaced by
`useNagivate` as this requires updating a _lot_ of places.

Several old class components has to be converted to functional
components to remove dependencies on the removed `withRouter`.

There's one outstanding problem: `history.block` is no longer available.
This is only used in a single place (`Form.tsx`), but it's a really
important place.