Filter unique authors when creating dialogue post fixer
We don't currently ensure that the list of authors only contains unique
authors when creating the dialogue post fixer. Having duplicate authors
is unlikely, but it is physically possible by adding the main post
author as a coauthor as well (this happened in testing as it isn't clear
from the "New Dialogue" dialog that you don't need to add yourself as an
author to be involved). Unfortunately, this causes an infinite loop and
crashes the browser tab which is... bad.