Update header "unFixed" value on page navigation
When we move to a different page we will be positioned at the top of
the page (unless the hash is set). Unfortunately Headroom doesn't run
the callback to update the value of "unfixed" on navigation so the
header thinks it is still scrolled. We can update the value manually
instead.