Add current user as a participant to conversations when messaging
If a moderator sends a message in a conversation they're not part of
we need to add them as a particpant in that conversation otherwise
they won't be notified of future messages.

We also have a notification for users when they are added to a
conversation. We need to change this to not send a notification to the
user if they added themselves.