Integrate `logIfSlow` directly into the `SqlClient` interface
Currently, you can wrap database calls in the `logIfSlow` function
which will print a warning message to the console if the query takes
longer than a pre-configured amount of time, along with an optional
description.

This PR bakes this functionality directly into the sql client, so
instead of: `logIfSlow(() => db.any("", []), "my-slow-query")` you can
now simply `db.any("", [], "my-slow-query")`. Slow-query logging is
applied to _all_ DB calls made with the client, whether or not they
supply a description - if no description if provided we just use the SQL
string (we don't log the arguments as they may contain sensitive data).

We also add better typing for the `args` parameter of all sql functions.