UsersProfile fragment no longer inherits SunshineUsersList fragment
The UsersProfile fragment was inheriting the SunshineUsersList fragment,
which includes some queries that are really only intended for the
moderator sidebar, but which, by being in the UsersProfile fragment,
would also be in the UsersCurrent fragment. This slows down SSR because
everything in UsersCurrent has to finish resolving before page rendering
can proceed.