Key `CollectionBase` off of `CollectionNameString` instead of `DbObject`
This is a very large and intrusive change, but it's worth it as it
tidies up a lot of edge cases in our type system.

If we have a `CollectionNameString` we can convert it into the matching
`DbObject` using `ObjectsByCollectionName`, but there's no way to do the
reverse operation as an object type can't be a key in a Typescript
interface. For this reason, `CollectionNameString` can be thought off as
the canonical "base" of a collection, and it makes it much easier to
correctly type things if we define it as such in the code.

This means that you will now write `BaseCollection<"Posts">` or
`PgCollection<"Comments">` instead of the old `BaseCollection<DbPost>`
or `PgCollection<DbComment>`. Generic functions with these arguments
will now define their type parameter as `<N extends CollectionNameString>`
instead of `<T extends DbObject>`.

Many other related types and functions are also now keyed off of
`CollectionNameString` to match (including but not limited to
`WithUpdateFunction`, `WithCreateFunction`, `SchemaType` and
`CollectionFieldSpecification`).

As part of the conversion, I also tidied up a few typing issues. Most
notably, `accessFilterSingle` and `accessFilterMultiple` are now typed
as returning partial DB objects rather than full DB objects. This still
isn't perfectly true, but it's more true than the previous
implementation and doing it "correctly" is very non-trivial.

One downside of this change is that it's now slightly harder to define
subsets of collections, such as the subset that have a `createdAt` field
or the subset that have a `slug` field. My approach to fixing this is to
simply add more generated types; in this case `CollectionNameWithCreatedAt`
and `CollectionNameWithSlug`.

Whilst touching some very old files, I also took the opportunity to
remove some dead-code and to change some `DEPRECATED` comments into
proper `@deprecated` doc-comments.