Fix post actions button being broken under react 18
Something in React 18 changed the timing of mouse event dispatch, which
made it so that when you clicked on the triple-dot menu on posts (and
possibly on other things), the click-away listener would trigger
immediately, making the button effectively inoperable.

ClickAwayListener was from material-UI, which is fairly out of date,
(see https://github.com/mui/material-ui/issues/38081 which is related
but not quite the same bug). Rather than take on upgrade material-UI,
vendor `react-click-away-listener` (which is MIT licensed), make two
unwrapped usages of `ClickAwayListener` point to the existing wrapper
`LWClickAwayListener`, and make `LWClickAwayListener` use the vendored
library.