Speed up truncatise()
Truncatise is a library which takes HTML, and shortens it to a target
character, word, or paragraph count. We vendored/forked this library
long ago. It's not very fast; it's written as-if a Shlemiel the Painter
algorithm (though I think V8 sneakily optimizes that away), and does a
bunch of unnecessary memory-allocating and regexing. Rewrite it,
simplifying the code, dropping the unused StripHTML feature, and greatly
speeding it up. Performance improvement for the logged-out front page
was about 30ms (5%) overall (using the profiler to look only at
the highlightFromHTML and truncatise functions). Also add a few unit
tests for it.