Maintain scroll position when changing comment sort order
Prior to this change, the navigation that occurs when you change the
comment sort order would either:
1) Re-navigate to the permalinked comment, if you're on a comment
permalink page, or
2) Scroll to the top.

This change maintains the current scroll position when the sort order is
changed. This required a little finessing; when the comment sort order
is changed, the comment list section has a zero height while the
comments reload, which could jolt the scroll position up, so now we
track whether comments are loading and add a minimum height while they
are.

Also, LW and EAForum put permalinked comments at the top of the
page. When changing the sort order on such a page, we change to a
regular post page, removing this top element, which would cause an
apparent jolt to the scroll position. So we record the height of that
element and adjust the scroll position by that amount.