Don't fetch the current user for the /api/search endpoint
We currently fetch the current user when making calls to the search
endpoint. This is a _blocking_ call - we can't even begin building the
search query until we get the result back from Postgres, let alone send
that query to elastic. However, we never actually use the current user
in these search queries (permissions are handled when exporting to
elastic) so we can safely remove this fetch to make searching feel a bit
snappier.