Spread out duplicate map locations
On the community map, many people enter imprecise locations like a city
name. These get displayed exactly on top of each other, which makes it
so that only one of them can be seen. Change that so that when multiple
map markers would be in the exact same location, we spread them out a
bit.