Fix custom filtering from frontpage filter settings
We use the filters to create a synthetic field to use for sorting called “score”
which shadows the “score” field that exists in the database. When we recently
implemented joins in the query builder we had to start qualifying the table name
for fields in expressions, so `ORDER BY "score" DESC` became
`ORDER BY "p"."score" DESC` which means we’re now sorting based on the actual
field in the database instead of the synthetic field that it shadows.
The simplest fix is to rename the synthetic field to avoid shadowing.