Add loading-state placeholders for post items, and tweak heights so the front page loading state lines up
Makes it so that if you use the service worker preload for the front
page, you get a correctly aligned skeleton of the front page. Tweaks the
height of LWPostsItem for bookmarked posts, shrinking it ~1px to make it
the same as regular post items, to prevent a mismatch. Removes a stray
bottom border from RecommendationsList. Adds a bookmarkCount field to
users, to be used to render the bookmarks list on the front page without
having to wait for the bookmarked posts.

Note that the spotlight item is _not_ handled, so if you haven't
dismissed the current spotlight item, it will shift everything down when
it appears. If you have dismissed it, though, this looks great.