Remove treeShaking: "ignore-annotations" from build.js
This option was added to our build script a long time ago to work around
a problem with Algolia, which apparently had tree-shaking annotations in
it that were incorrect. However, we are mostly not relying on Algolia
anymore. Removing this annotation reduces the (minified, uncompressed)
client bundle size by about 2%. We retested the client-side search UI
and confirmed that it still works with this change.

Co-Authored-By: Ollie Etherington <ollie@etherington.xyz>