Speed up side-comment matching
Speeds up `matchSideComments` by early-exiting on comments with no
quote-shards, and hoisting a lot of the computation out of the
per-comment loop. On the post
  https://www.lesswrong.com/posts/wAczufCpMdaamF9fy/my-objections-to-we-re-all-gonna-die-with-eliezer-yudkowsky
tested on my development laptop, the part of the pageload attributed to this
function goes from ~430ms to ~120ms.

This function being slow was having a noticeable impact because
side-comment caching wasn't working, possibly due to the new
fragment-compilation code. That's fixed in
https://github.com/ForumMagnum/ForumMagnum/pull/8769 which splits the
side-comments cache into its own table, so this isn't as necessary as it
would be without that, but it'll still help with cases like when the
cache gets populated for the first time (typically by a crawler visiting
old posts) and for pageloads after comment submission.