Upgrade version of postgres-array
Postgres-array is a dependency of pg-types which is a depndency of pg.
It's responsible for parsing jsonb fields in the results of queries we
run in postgres. This version upgrade causes us to get the changes in
https://github.com/bendrucker/postgres-array/pull/16 , which makes that
particular function ~2-3x faster. I observed that this function was ~2%
of our server CPU time (on a benchmark where I was SSRing a post page
with lots of comments in a loop), and after this upgrade I saw the time
spent in it reduce accordingly.

The version is a major version only because they dropped compatibility
with some old nodejs versions.