Use linkedom instead of JSDOM for ToC
In extractTableOfContents, we parse a post and find headings in it.
Previously, this was done with cheerio. Recently it was modified to
JSDOM so that the same code could run on the client (though in practice
it's usually done server side). Unfortunately this was slower,
especially on large posts; on an 11k-word real post, it was adding
~80ms of CPU time to per SSR (on my M1 Max).

There's a competing library linkedom which is a lot faster, with a
similar API. Switch libraries, and smooth over a few small API
differences.