Fix `getTagBotUserId` retaining and leaking a ResolverContext
The performance impact of this one isn't super high, but the pattern is
interesting to watch out for.

The first time `getTagBotUserId` is called, it fetches a user by slug
and returns its ID; on subsequent calls, it reuses the result of that
fetch and returns the same ID. In the course of doing this, it creates a
promise and stores it in a global variable (so subsequent calls can wait
for the first call to finish, rather than starting their own DB
fetches).

After everything is settled, that promise was retained in a global
variable. It turns out, retaining the promise (even in a resolved state)
retains everything captured by the promise's execution context, which
means that the first call to this function takes its arguments (a
ResolverContext) and leaks them. (When I found it in a heap snapshot, the
debugger claimed it was retaining 23MB. This amount of memory is not a
big deal, but there may be contexts in which the amount leaked much
larger.)