Remove minifyAsIfProduction option from build.js
There was (in the bundle-split-react-map-gl branch only) a flag
minifyAsIfProduction, which turns on minification without making the
bundle use production configs in other ways. This was for measuring the
effect of bundle splitting. But it turns out it causes the `graphql`
library to crash on load, in a way that breaks all other JS on the page.
Specifically, as a guard against having multiple copies of itself linked
into the same page, it checks whether (a) it's in development mode and
(b) its internals have been renamed. It assumes that if it finds pieces
of itself were renamed in production mode, it's due to a minifier and
can be ignored, but if pieces of itself are renamed in development mode,
it's due to a bundler resolving name collisions, and indicates a
problem. This is a false positive.