Avoid making a list of a user's group-memberships for every field when permission filtering
Previously when checking field permissions on objects, we were calling
userIsMemberOf for each permissions-list entry of each field of each
filtered object; and userIsMemberOf calls userGetGroups, which
constructs a new list each time. Reduce that so that the list of groups
is only generated once per object, speeding up
`restrictViewableFieldsSingle` and similar functions.