Replace Picker library with something more straightforward
When looking at why server-side async context switches were slow, I
found a lot of time spent in `async_hooks`; this appears to correspond
to something that Sentry, Datadog, and possibly some unidentified third
library added so that they can track stack traces across async/await
boundaries. I also noticed that, by the time we get to renderPage, the
call stack is about 200 frames deep. That seems like it might have
something to do with async_hooks callbacks being slow.

Most of those stack frames correspond to Express middlewares, and the
single largest contributor is a library called Picker, which does
routing for static routes. Nearly all its functionality is unused, and
the reason it adds a lot of stack frames is because it unsuccessfully
tries to do tail recursion. So, simplify it down to only the
functionality we actually use, merge it into `staticRoutes.ts`, and
rewrite the tail-recursive part to be iterative.

(I preserved the copyright notice and MIT license that was in the
vendored library, even though the result bears little resemblance to the
original vendored library, because I generated it incrementally.)