Allow admins to read value of `hideSunshineSidebar`
Currently, when an admin tries to edit a user other than themselves they
get an error when they submit. This is because they don't have
permission to read `hideSunshineSidebar`, so they inadvertantly try to
assign a value of `null` to a field with a default value which is not
allowed.