Merge in-flight requests in the DB cache
When two different instances get logged-out requests for the same page,
and the second request arrives while the first one is still rendering,
make the second detect that there's an in-progress render via the page
cache and wait for it.

Theoretically this should work, but it still needs more testing/debugging.

Co-Authored-By: Ollie Etherington <ollie@etherington.xyz>