Upgrade CkEditor package versions; start enabling Typescript for CkEditor plugins
This upgrades the package versions of CkEditor and all its first-party
plugins to latest, and deals with some webpack config and ckeditor
plugin config changes this requires. It also enables Typescript for
CkEditor plugins (but doesn't yet make tsc autorun on the plugins the
way it does for the main repo, and doesn't yet pull the trigger on
converting any particular plugins).