Unsetting `submitToFrontpage` unsets `frontpageDate`
Currently, if you publish a post with `submitToFrontpage` set to true
then the post will appear on the frontpage due to the fact that
`frontpageDate` is set to the current time. However, this check is only
done at the time that the post is _published_, and if you later toggle
`submitToFrontpage` to false the post will still appear on the frontpage
because `frontpageDate` will still be set. This change fixes the issue
by unsetting `frontpageDate` whenever `submitToFrontpage` is unset.