Fix error in PresenceList when logged-out (link-sharing) users are connected
When link sharing is enabled, logged-out users can connect to
collaborative editing sessions. When this happens, for CkEditor cloud
purposes, they get a random ID in place of a user ID. In the presence
list, we were taking that random ID and trying to look it up as though
it was a user ID, resulting in a lot of console errors, and anonymous
users not being shown in the presence-list indicator.

This change handles anon users in the presence list properly; they now
show up in the presence list when connected, and don't produce
console/Sentry errors.