Configure typescript, eslint and jest to allow absolute imports
Currently, any time we import a file, we import it with a path that's
relative to the file that's doing the importing. Sometimes this means
imports that look like

    import { foo } from "../../../../lib/foo"

Which is pretty silly, and adds some extra friction any time we want to
move files around. This commit adds "paths" to tsconfig.json, and
corresponding config settings for eslint and jest, so you can instead
write

    import { foo } from "@/lib/foo"

Any paths that start with @/ are interpreted as being relative to
packages/lesswrong. Because this is a first-class feature of Typescript
(we just needed to enable it in tsconfig.json), Typescript language
servers seem to already support it with no additional configuring.
Auto-imports are still likely to give you paths with ".."s in them,
though.

To make this work with eslint, we add the plugin
eslint-import-resolver-typescript. This causes a bunch of
eslint-plugin-import rules, which were apparently not working, to wake
up and start checking things. Unfortunately a lot of that checking is
(a) redundant with what's built into Typescript, and (b) subject to
false positives. So we turn off import/default, import/namespace, and
import/named, rules which get confused by a few specific libraries in
node_modules. We also turn off import/no-deprecated, because the way
we're importing cheerio is in fact deprecated, but fixing it is a bit
more involved than just changing import statements so I left a comment
to do it later.