Fix bundle/stylesheet preload for AF-to-LW crossposts
When we start loading a post page, before we start rendering it in an
SSR, we sometimes send the part of the header that tells the client to
start downloading the stylesheet and javascript bundle. Before we can
send that, we need to be sure the request is going to result in an http
200 (ie, not a 404 or a redirect). This is handled by
`PostsRepo.postRouteWillDefinitelyReturn200`. However, this handled the
AF flag incorrectly; rather than checking whether the user is loading a
non-AF post on AF (which would trigger a redirect), it checks whether
the user is loading an AF post (which can happen on either LW or AF, and
will not trigger a redirect). This meant that AF posts didn't get the
bundle/stylesheet preload, making them slower.