Only initialize the CkEditor versions that will be used
The way wrapCkEditor was written, you have to import (and trigger some
plugin initialization for) every version of the editor (comment, post,
and collaborative), regardless of which you're going to use.

Make it so that you only set up the config for the one you're going to
use, and memoize the config initialization.