Reduce use of collections, first pass
In general, components can use and pass around collection objects, or
collection names. We'd rather they do the latter. Eventually, the goal
here is that we can finish hydration after without importing any
collection schemas, and maybe get the schemas out of the bundle
entirely.

As a first pass, when getCollection() is used just for conversions
between collectionName and typeName, use a consistent wrapper for that;
switch some places that are using collections when they could be using
collection names, to using only names; and remove unused imports of
collection files.