Make the permissions-lists in user groups immutable
Rather than creating groups and then calling .can() to add permissions
to them, the full permissions list is provided to the createGroup()
call. This means moving all the permissions from individual
collection/permissions.ts files, to lib/permissions.ts.

(Putting all the permissions next to each other is making it easy to
notice how dumb and broken a lot of it is.)