Some links navigate on press, rather than on release
See: https://x.com/ID_AA_Carmack/status/1787850053912064005

Adds a prop `doOnDown` to our `Link` component which, if true, makes the
link navigate on press rather than on release. Currently we apply this
only to `PostsTitle` (which is used in `PostsItem`). Also refactor the
`Link` component so that the `to` field is always a string, replacing a
few cases where it was an object and was relying on `react-router-dom`
to do string-munging to turn it into a link.

If feedback on this is positive, we may wish to change the default value
of `doOnDown` to `true`, so that it applies to all links, not just post
links. For now we're not doing that because there's some risk that some
weird link we're not paying attention to interacts badly with it, eg a
link with a custom `onClick` handler.

Co-Authored-By: Robert M <bobert.mushky@gmail.com>