Fix upcast of `footnoteItem`
When we upgraded CkEditor to the latest version, some subtle load order
change (or similar) made it so that the data upcast of `footnoteSection`
didn't work. The issue is that this element, in the data representation,
is an `<ol>`, and the ListProperties plugin also has an upcast for
`<ol>` elements, and it took precedence. We fix this by adding
`converterPriority: 'high'` to our upcast, resolving the ambiguity.

The symptom of this issue is that when you saved a post that contained
footnotes and then reopened it, the footnotes would move out of the
footnote block to the end of the post.