Make collapsible sections use elementToStructure
This makes it so that rather than the title and content blocks being
special element types in the schema, the title is a regular element in a
slot, which is defined positionally. This prevents a lot of broken
states that the collapsible sections were able to end up in, which
improves the UI greatly.