Fix memory leak on post pages
Our T3AudioPlayer component is responsible for loading an external
script from embed.type.audio. As written, it would delete and re-add
this script tag whenever you navigated to and from post pages, meaning
the external script's initialization would run repeatedly, leaking
memory each time. The practical upshot of that was that if you read the
site by navigating to post pages and then using the back button, the tab
would leak 5MB of memory (as measured by the Chrome devtools memory
profiler) for each post you looked at.

Change the script loading so that once the T3 embed script is added to
the page, it stays there and is never removed. This fixes the memory
leak.

I tested on localhost by using a hack that intercepts the `fetch`
function and replaces the localhost domain in type3 API calls with
lesswrong.com. I tested that you can land on an SSR'ed post page and
then use the player (there is some loading-flicker if the player starts
out expanded but that was already there), that you can start on a
non-post page, navigate to a post, and use the player, and that you can
navigate to multiple posts and the player still works after the first
one.