[WIP] Use full-height ToC on the edit-post page
Makes the edit-post page use the full-height ToC (if that's enabled on
post pages). This turns out to be more architecturally-involved than
you'd think, because the ToC code was built around the assumption that
you can run a preprocessing step over the post body HTML, to give things
IDs to use as anchors. But you can't futz with the contents of CkEditor
like that. With the old ToC code, we could skip the HTML substitution at
the expense of not being able to highlight the active section or make
the section links clickable. With the full-height ToC, the list of
features that don't work under that architecture expands to also include
"dynamic vertical sizing", which is one missing feature too many.

So, we modify `extractTableOfContents` to take an `includeElementRefs`
option which, if passed, causes ToC sections to include a reference to
their heading element (to be used instead of finding that element by
it's ID attribute, since it doesn't have an ID attribute). Then we make
the `DynamicTableOfContents` component use DOM element that it finds in
the browser page, rather than serialized HTML, so that it can get those
references.

Known issues:
 * FixedPositionToC does its height detection, plus some
   section-heading-filtering,  in a `useEffect` hook on startup, which
   interferes with `DynamicTableOfContents` trying to update it to
   reflect edits made in the post.
 * Also, `DynamicTableOfContents` isn't set up to trigger updates
   properly either. It needs a debounced timer and some related things.
 * `extractTableOfContents` isn't pure; it actually does the
   ToC-extraction and the HTML-preprocessing in one step. When called
   from `useEditorTableOfContents`, it side-effectfully assigns ID
   attributes to elements it finds, which is not kosher when CkEditor
   owns those elements.
 * Some ToC features, like current-section highlighting, still rely on
   anchors. Those features are currently working, but will stop working
   when the previous issue is addressed.
 * The ToC (on the post-edit page) has a comment-section link at the
   bottom that doesn't belong there and doesn't go anywhere.
 * The new-post and collaboration-only editor pages have not been
   updated at all, only the edit-post page